fix bug and add preview func

+++ route/board_edit.js
@@ -12,38 +12,45 @@ function board_edit(req, res) {
                 req.params.b_name,
                 req.params.id
             ], function(err, db_data) {
-                db.all("select set_data from user_data where user_name = ? and set_name = 'auth'", [user_name], function(err, db_data_2) {
-                    if(db_data_2[0].set_data === 'admin' || req.session['user_name'] === db_data[0].doc_data) {
-                        let data = req.body;
+                if(db_data.length > 0) {
+                    db.all("select set_data from user_data where user_name = ? and set_name = 'auth'", [user_name], function(err, db_data_2) {
+                        if(db_data_2[0].set_data === 'admin' || req.session['user_name'] === db_data[0].doc_data) {
+                            let data = req.body;
 
-                        let title = data.title;
-                        let content = data.content;
+                            let title = data.title;
+                            let content = data.content;
 
-                        if(content === '') {
-                            db.run("delete from bbs_data where doc_id = ?", [req.params.id]);
+                            if(content === '') {
+                                db.run("delete from bbs_data where doc_id = ?", [req.params.id]);
+                            } else {
+                                db.run("update bbs_data set doc_data = ? where set_data = ? and doc_id = ? and set_name = 'title'", [
+                                    title,
+                                    req.params.b_name,
+                                    req.params.id
+                                ]);
+                                db.run("update bbs_data set doc_data = ? where set_data = ? and doc_id = ? and set_name = 'content'", [
+                                    content,
+                                    req.params.b_name,
+                                    req.params.id
+                                ]);
+                            }
+
+                            res.json({
+                                "req" : "ok"
+                            });
                         } else {
-                            db.run("update bbs_data set doc_data = ? where set_data = ? and doc_id = ? and set_name = 'title'", [
-                                title,
-                                req.params.b_name,
-                                req.params.id
-                            ]);
-                            db.run("update bbs_data set doc_data = ? where set_data = ? and doc_id = ? and set_name = 'content'", [
-                                content,
-                                req.params.b_name,
-                                req.params.id
-                            ]);
+                            res.json({
+                                "req" : "error",
+                                "reason" : "user_name !== doc_user_name && user_auth !== \"admin\""
+                            });
                         }
-
-                        res.json({
-                            "req" : "ok"
-                        });
-                    } else {
-                        res.json({
-                            "req" : "error",
-                            "reason" : "user_name !== doc_user_name && user_auth !== \"admin\""
-                        });
-                    }
-                });
+                    });
+                } else {
+                    res.json({
+                        "req" : "error",
+                        "reason" : "document not exist"
+                    });
+                }
             });
         } else {
             res.json({
